Move `ConnectionMatcher` to ConnectionMatcher.cs
